BROKEN: trying to track down a weird error:
PHP Fatal error:  Uncaught exception 'PharException' with message 'unable to copy stub of old phar to new phar "/vagrant/puppet/modules/vagrant_sync/daemon/build/vagrant_sync.phar"' in /vagrant/puppet/modules/vagrant_sync/daemon/source/createPhar.php:49
Stack trace:
\#0 /vagrant/puppet/modules/vagrant_sync/daemon/source/createPhar.php(49): Phar->offsetSet('namespace/Appli...', '<?php?namespace...')
\#1 {main}
  thrown in /vagrant/puppet/modules/vagrant_sync/daemon/source/createPhar.php on line 49

+++ daemon/source/createPhar.php
@@ -1,11 +1,38 @@
 <?php
 
+function recursiveFileAdd(&$filesToAdd, $startPath, $fromPath, $excludeList)
+{
+	$startPath = rtrim($startPath, '/\\');
+	if (!file_exists($fromPath)) {
+		throw new Exception('Directory doesn\'t exist.');
+	} elseif (!is_dir($fromPath)) {
+		$finalFileName = substr($fromPath, (strlen($startPath) + 1));
+		if (!in_array($finalFileName, $excludeList) ) {
+			$filesToAdd[$finalFileName] = $fromPath;
+		}
+	} else {
+		$directoryIterator = new DirectoryIterator($fromPath);
+
+		foreach ($directoryIterator as $fileInfo) {
+			if (!$fileInfo->isDot()) {
+				recursiveFileAdd($filesToAdd, $startPath, $fileInfo->getPathname(), $excludeList);
+			}
+		}
+	}
+}
+
 $srcRoot = dirname(__FILE__);
 $buildRoot = dirname($srcRoot) . DIRECTORY_SEPARATOR . "build";
 $archiveBaseName = "vagrant_sync";
 $archiveName = "{$archiveBaseName}.phar";
 $archivePath = $buildRoot . DIRECTORY_SEPARATOR . $archiveName;
 
+if (ini_get('phar.readonly')) {
+	echo 'phar.readonly must be set to false before running this script!';
+	exit(2);
+}
+$filesToAdd = array();
+recursiveFileAdd($filesToAdd, $srcRoot, $srcRoot, array('stub.php', 'createPhar.php'));
 if (file_exists($archivePath)) {
 	if (!unlink($archivePath)) {
 		throw new Exception("Unable to delete file $archivePath");
@@ -16,7 +43,10 @@ $phar = new Phar(
 	(FilesystemIterator::CURRENT_AS_FILEINFO | FilesystemIterator::KEY_AS_FILENAME),
 	$archiveName
 );
-$phar["autoloader.php"] = file_get_contents($srcRoot . DIRECTORY_SEPARATOR . "autoloader.php");
 $phar->setStub(file_get_contents($srcRoot . DIRECTORY_SEPARATOR . "stub.php"));
+foreach ($filesToAdd as $filename => $fullPath) {
+	echo "Adding {$fullPath} as '{$filename}'...\n";
+	$phar[$filename] = file_get_contents($fullPath);
+}
 
 chmod($archivePath, 0755);
